--- a/akka-remote-tests/src/multi-jvm/scala/akka/remote/artery/LatencySpec.scala
+++ b/akka-remote-tests/src/multi-jvm/scala/akka/remote/artery/LatencySpec.scala
@@ -83,11 +83,11 @@ object LatencySpec extends MultiNodeConfig {
   }
 
   def receiverProps(reporter: RateReporter, settings: TestSettings, totalMessages: Int,
-                    sendTimes: AtomicLongArray, histogram: Histogram, plotsRef: ActorRef): Props =
-    Props(new Receiver(reporter, settings, totalMessages, sendTimes, histogram, plotsRef))
+                    sendTimes: AtomicLongArray, histogram: Histogram, plotsRef: ActorRef, BenchmarkFileReporter: BenchmarkFileReporter): Props =
+    Props(new Receiver(reporter, settings, totalMessages, sendTimes, histogram, plotsRef, BenchmarkFileReporter))
 
   class Receiver(reporter: RateReporter, settings: TestSettings, totalMessages: Int,
-                 sendTimes: AtomicLongArray, histogram: Histogram, plotsRef: ActorRef) extends Actor {
+                 sendTimes: AtomicLongArray, histogram: Histogram, plotsRef: ActorRef, BenchmarkFileReporter: BenchmarkFileReporter) extends Actor {
     import settings._
 
     var count = 0
@@ -122,12 +122,12 @@ object LatencySpec extends MultiNodeConfig {
           }
       }
       if (count == totalMessages) {
-        printTotal(testName, size, histogram, System.nanoTime() - startTime)
+        printTotal(testName, size, histogram, System.nanoTime() - startTime, BenchmarkFileReporter)
         context.stop(self)
       }
     }
 
-    def printTotal(testName: String, payloadSize: Long, histogram: Histogram, totalDurationNanos: Long): Unit = {
+    def printTotal(testName: String, payloadSize: Long, histogram: Histogram, totalDurationNanos: Long, reporter: BenchmarkFileReporter): Unit = {
       import scala.collection.JavaConverters._
       val percentiles = histogram.percentiles(5)
       def percentile(p: Double): Double =
@@ -138,7 +138,7 @@ object LatencySpec extends MultiNodeConfig {
 
       val throughput = 1000.0 * histogram.getTotalCount / math.max(1, totalDurationNanos.nanos.toMillis)
 
-      println(s"=== Latency $testName: RTT " +
+      reporter.reportResults(s"=== Latency $testName: RTT " +
         f"50%%ile: ${percentile(50.0)}%.0f µs, " +
         f"90%%ile: ${percentile(90.0)}%.0f µs, " +
         f"99%%ile: ${percentile(99.0)}%.0f µs, " +
@@ -244,7 +244,7 @@ abstract class LatencySpec
       repeat = repeatCount,
       realMessage))
 
-  def test(testSettings: TestSettings): Unit = {
+  def test(testSettings: TestSettings, BenchmarkFileReporter: BenchmarkFileReporter): Unit = {
     import testSettings._
 
     runOn(first) {
@@ -271,7 +271,7 @@ abstract class LatencySpec
         echo ! Reset
         expectMsg(Reset)
         histogram.reset()
-        val receiver = system.actorOf(receiverProps(rep, testSettings, totalMessages, sendTimes, histogram, plotProbe.ref))
+        val receiver = system.actorOf(receiverProps(rep, testSettings, totalMessages, sendTimes, histogram, plotProbe.ref, BenchmarkFileReporter))
 
         // warmup for 3 seconds to init compression
         val warmup = Source(1 to 30)
@@ -335,6 +335,7 @@ abstract class LatencySpec
   }
 
   "Latency of Artery" must {
+    val reporter = BenchmarkFileReporter("LatencySpec", system)
 
     "start echo" in {
       runOn(second) {
@@ -345,7 +346,7 @@ abstract class LatencySpec
     }
 
     for (s ← scenarios) {
-      s"be low for ${s.testName}, at ${s.messageRate} msg/s, payloadSize = ${s.payloadSize}" in test(s)
+      s"be low for ${s.testName}, at ${s.messageRate} msg/s, payloadSize = ${s.payloadSize}" in test(s, reporter)
     }
 
     // TODO add more tests