@jrudolph 2016-12-30
Fixed all the mentioned issues and implemented the new strategy as suggested by @patriknw:

    For inbound-lanes > 1 or idle-cpu-level < 5   : no spinning in the graph stage
    For inbound-lanes = 1 and idle-cpu-level >= 6 : 50 * idle-cpu-level - 240