@patriknw 2017-01-02
I have started the https://jenkins.akka.io:8498/job/akka-artery-repeat/ job for this branch. Let's run it a few times to get a preview of how it behaves on jenkins