@jrudolph 2017-01-02
@patriknw, thanks for starting the tests.